+++ speechtools/gui/widgets/query/__init__.py
@@ -1,0 +1,2 @@
+
+from .main import QueryWidget
