+++ speechtools/gui/widgets/query/main.py
@@ -5,229 +5,17 @@ from speechtools.corpus import CorpusContext
 
 from polyglotdb.graph.func import Sum, Count
 
-from .base import DetailedMessageBox
+from ..base import DetailedMessageBox
 
-from ..models import QueryResultsModel, ProxyModel
+from ...models import QueryResultsModel, ProxyModel
 
-from ..views import ResultsView
+from ...views import ResultsView
 
-from ..workers import (QueryWorker, Lab1QueryWorker, ExportQueryWorker)
+from ...workers import (QueryWorker, Lab1QueryWorker, ExportQueryWorker)
 
-class NonScrollingComboBox(QtWidgets.QComboBox):
-    def __init__(self, parent = None):
-        super(NonScrollingComboBox, self).__init__(parent)
-        self.setFocusPolicy(QtCore.Qt.StrongFocus)
+from .graphical import GraphicalQuery
 
-    def wheelEvent(self, e):
-        e.ignore()
-
-class AnnotationRect(QtWidgets.QGraphicsTextItem):
-    clicked = QtCore.pyqtSignal(object)
-    def __init__(self, text, hierarchy, pos = 0):
-        super(AnnotationRect, self).__init__(text)
-        self.linear_position = pos
-        self.hierarchy = hierarchy
-        for i, at in enumerate(self.hierarchy.lowest_to_highest):
-            if at == text:
-                self.hierarchy_position = i
-                break
-
-    def paint(self, painter, option, widget = 0):
-        painter.setPen(QtCore.Qt.black)
-        painter.setBrush(QtGui.QColor(230,230,230))
-        r = option.rect
-        mod = 50 * self.hierarchy_position
-        r.setWidth(r.width() + mod)
-        r.setX(0 - r.width() / 2)
-        painter.drawRect(r)
-        super(AnnotationRect, self).paint(painter, option, widget)
-
-    def mousePressEvent(self, event):
-        super(AnnotationRect, self).mousePressEvent(event)
-        if event.button() == QtCore.Qt.LeftButton:
-            self.clicked.emit(self.toPlainText())
-
-class FilterWidgetItem(QtWidgets.QWidget):
-    def __init__(self, column, value):
-        super(FilterWidgetItem, self).__init__()
-
-        self.column = column
-        self.value = value
-
-        layout = QtWidgets.QHBoxLayout()
-        layout.addWidget(QtWidgets.QLabel(column))
-
-        if value in (int, float):
-            self.compWidget = NonScrollingComboBox()
-            self.compWidget.addItem('=')
-            self.compWidget.addItem('!=')
-            self.compWidget.addItem('>')
-            self.compWidget.addItem('>=')
-            self.compWidget.addItem('<')
-            self.compWidget.addItem('<=')
-            layout.addWidget(self.compWidget)
-            self.valueWidget = QtWidgets.QLineEdit()
-            layout.addWidget(self.valueWidget)
-        self.setLayout(layout)
-
-class SubsetWidgetItem(QtWidgets.QWidget):
-    def __init__(self):
-        super(SubsetWidgetItem, self).__init__()
-
-        self.subsetList = NonScrollingComboBox()
-        self.subsetList.addItem('')
-
-        layout = QtWidgets.QHBoxLayout()
-        layout.addWidget(QtWidgets.QLabel('Subset'))
-        layout.addWidget(self.subsetList)
-        self.setLayout(layout)
-
-    def clear(self):
-        self.subsetList.clear()
-        self.subsetList.addItem('')
-
-    def addItems(self, items):
-        for i in items:
-            self.subsetList.addItem(i)
-
-class FilterWidget(QtWidgets.QGroupBox):
-    def __init__(self):
-        self.hierarchy = None
-        super(FilterWidget, self).__init__('Filters')
-
-        mainlayout = QtWidgets.QVBoxLayout()
-
-        self.filterLayout = QtWidgets.QVBoxLayout()
-
-        self.subsetWidget = SubsetWidgetItem()
-
-        self.filterLayout.addWidget(self.subsetWidget)
-
-        self.attributeBox = QtWidgets.QGroupBox('Attributes')
-
-        self.attributeBox.setLayout(self.filterLayout)
-
-        mainlayout.addWidget(self.attributeBox)
-
-        self.speakerBox = QtWidgets.QGroupBox('Speakers')
-
-        mainlayout.addWidget(self.speakerBox)
-        self.discourseBox = QtWidgets.QGroupBox('Discourses')
-
-        mainlayout.addWidget(self.discourseBox)
-
-        mainWidget = QtWidgets.QWidget()
-        mainWidget.setLayout(mainlayout)
-
-        layout = QtWidgets.QVBoxLayout()
-        scroll = QtWidgets.QScrollArea()
-        scroll.setWidgetResizable(True)
-        scroll.setWidget(mainWidget)
-        scroll.setMinimumHeight(200)
-        policy = scroll.sizePolicy()
-        policy.setVerticalStretch(1)
-        scroll.setSizePolicy(policy)
-        layout.addWidget(scroll)
-
-        self.acceptButton = QtWidgets.QPushButton('Add filters')
-
-        layout.addWidget(self.acceptButton)
-
-        self.setLayout(layout)
-
-    def setHierarchy(self, hierarchy):
-        self.hierarchy = hierarchy
-
-    def updateAnnotation(self, text):
-        while self.filterLayout.count():
-            item = self.filterLayout.takeAt(0)
-            if item.widget() is None:
-                continue
-            item.widget().deleteLater()
-
-        self.subsetWidget = SubsetWidgetItem()
-
-        self.filterLayout.addWidget(self.subsetWidget)
-        subsets = []
-        if text in self.hierarchy.subset_tokens:
-            for s in self.hierarchy.subset_tokens[text]:
-                subsets.append(s)
-        if text in self.hierarchy.subset_types:
-            for s in self.hierarchy.subset_types[text]:
-                subsets.append(s)
-        self.subsetWidget.addItems(subsets)
-        if text in self.hierarchy.type_properties:
-            for p, v in self.hierarchy.type_properties[text]:
-                self.filterLayout.addWidget(FilterWidgetItem(p, v))
-        if text in self.hierarchy.token_properties:
-            for p, v in self.hierarchy.token_properties[text]:
-                self.filterLayout.addWidget(FilterWidgetItem(p, v))
-
-class GraphicalQuery(QtWidgets.QWidget):
-    vert_step = 50
-    horz_step = 200
-    def __init__(self):
-        super(GraphicalQuery, self).__init__()
-        self.hierarchy = None
-        visualLayout = QtWidgets.QHBoxLayout()
-
-        self.filterWidget = FilterWidget()
-
-        self.queryScene = QtWidgets.QGraphicsScene()
-        self.queryView = QtWidgets.QGraphicsView(self.queryScene)
-        self.queryView.setContextMenuPolicy(QtCore.Qt.CustomContextMenu)
-        self.queryView.customContextMenuRequested.connect(self.showMenu)
-
-        visualLayout.addWidget(self.queryView)
-        visualLayout.addWidget(self.filterWidget)
-
-        self.setLayout(visualLayout)
-
-    def showMenu(self, pos):
-
-        menu = QtWidgets.QMenu()
-        items = self.queryView.items(pos)
-        if not items:
-            return
-        previous_action = QtWidgets.QAction('Add previous annotation', self)
-        following_action = QtWidgets.QAction('Add following annotation', self)
-
-        menu.addAction(previous_action)
-        menu.addAction(following_action)
-
-        action = menu.exec_(self.mapToGlobal(pos))
-        if action == previous_action:
-            new_pos = items[0].linear_position - 1
-        elif action == following_action:
-            new_pos = items[0].linear_position + 1
-        else:
-            return
-        self.addNew(items[0].toPlainText(), new_pos)
-
-    def addNew(self, text, pos):
-        for i, at in enumerate(self.hierarchy.highest_to_lowest):
-            if at == text:
-                break
-        ar = AnnotationRect(text, self.hierarchy)
-        ar.clicked.connect(self.filterWidget.updateAnnotation)
-        ar.setPos(self.horz_step * pos, self.vert_step*i)
-
-        br = ar.boundingRect()
-        br.setWidth(br.width() + self.vert_step*i)
-        self.queryScene.addItem(ar)
-
-    def setHierarchy(self, hierarchy):
-        self.hierarchy = hierarchy
-        self.filterWidget.setHierarchy(hierarchy)
-        self.queryScene.clear()
-
-        for i, at in enumerate(hierarchy.highest_to_lowest):
-            ar = AnnotationRect(at, hierarchy)
-            ar.clicked.connect(self.filterWidget.updateAnnotation)
-            ar.setPos(0, self.vert_step*i)
-            br = ar.boundingRect()
-            br.setWidth(br.width() + self.vert_step*i)
-            self.queryScene.addItem(ar)
+from .basic import BasicQuery
 
 class QueryForm(QtWidgets.QWidget):
     finishedRunning = QtCore.pyqtSignal(object)
@@ -238,13 +26,14 @@ class QueryForm(QtWidgets.QWidget):
         mainLayout = QtWidgets.QVBoxLayout()
         headerLayout = QtWidgets.QHBoxLayout()
 
-        self.graphicalQuery = GraphicalQuery()
+        self.graphicalQuery = BasicQuery()
 
 
         self.linguisticSelect = QtWidgets.QComboBox()
         self.querySelect = QtWidgets.QComboBox()
         self.querySelect.addItem('Lab 1 stops')
         self.querySelect.addItem('Lab 2 stops')
+        self.querySelect.addItem('Lab 3 stops')
         if not getattr(sys, 'frozen', False):
             mainLayout.addWidget(self.graphicalQuery)
             self.querySelect.addItem('Lab 1 laryngeal sampling')
@@ -296,11 +85,13 @@ class QueryForm(QtWidgets.QWidget):
             a_type = getattr(c, a_type)
             w_type = getattr(a_type, w_type)
             utt_type = getattr(a_type, utt_type)
-            if query_type in ['Lab 1 stops', 'Lab 2 stops']:
+            if query_type in ['Lab 1 stops', 'Lab 2 stops', 'Lab 3 stops']:
                 if query_type == 'Lab 1 stops':
                     filters.append(a_type.phon4lab1 == True)
                 elif query_type == 'Lab 2 stops':
                     filters.append(a_type.phon4lab2 == True)
+                elif query_type == 'Lab 3 stops':
+                    filters.append(a_type.phon4lab3 == True)
                 columns = [a_type.label.column_name('Stop'),
                             a_type.begin.column_name('Begin'),
                             a_type.end.column_name('End'),
@@ -317,6 +108,14 @@ class QueryForm(QtWidgets.QWidget):
                     columns.extend([a_type.closure.begin.column_name('Closure_begin'),
                                 a_type.closure.end.column_name('Closure_end'),
                                 Sum(a_type.closure.duration).column_name('Closure_duration')])
+                if 'preaspiration' in c.hierarchy.subannotations[c.hierarchy.lowest]:
+                    columns.extend([a_type.preaspiration.begin.column_name('Preaspiration_begin'),
+                                a_type.preaspiration.end.column_name('Preaspiration_end'),
+                                Sum(a_type.preaspiration.duration).column_name('Preaspiration_duration')])
+                if 'vowel_duration' in c.hierarchy.subannotations[c.hierarchy.lowest]:
+                    columns.extend([a_type.vowel_duration.begin.column_name('vowel_duration_begin'),
+                                a_type.vowel_duration.end.column_name('vowel_duration_end'),
+                                Sum(a_type.vowel_duration.duration).column_name('vowel_duration')])
 
                 columns.extend([w_type.label.column_name('Word'),
                             w_type.begin.column_name('Word_begin'),
@@ -361,11 +160,13 @@ class QueryForm(QtWidgets.QWidget):
             a_type = getattr(c, a_type)
             w_type = getattr(a_type, w_type)
             utt_type = getattr(w_type, utt_type)
-            if query_type in ['Lab 1 stops', 'Lab 2 stops']:
+            if query_type in ['Lab 1 stops', 'Lab 2 stops', 'Lab 3 stops']:
                 if query_type == 'Lab 1 stops':
                     filters.append(a_type.phon4lab1 == True)
                 elif query_type == 'Lab 2 stops':
                     filters.append(a_type.phon4lab2 == True)
+                elif query_type == 'Lab 3 stops':
+                    filters.append(a_type.phon4lab3 == True)
                 columns = [a_type.label.column_name('Stop'),
                             a_type.begin.column_name('Begin'),
                             a_type.end.column_name('End'),
@@ -389,6 +190,10 @@ class QueryForm(QtWidgets.QWidget):
                     columns.extend([a_type.closure.begin.column_name('Closure_begin'),
                                 a_type.closure.end.column_name('Closure_end'),
                                 Sum(a_type.closure.duration).column_name('Closure_duration')])
+                if 'vowel_duration' in c.hierarchy.subannotations[c.hierarchy.lowest]:
+                    columns.extend([a_type.vowel_duration.begin.column_name('vowel_duration_begin'),
+                                a_type.vowel_duration.end.column_name('vowel_duration_end'),
+                                Sum(a_type.vowel_duration.duration).column_name('vowel_duration')])
             else:
                 if query_type == 'Lab 1 laryngeal sampling':
                     filters.extend([a_type.begin == w_type.begin,
